moved kubectl delete test
