+++ tests/ha/single_pod_restart/test_server_pod_restart_api.py
@@ -927,169 +927,3 @@ class TestServerPodRestartAPI:
                                   f"Actual checksum: {download_checksum}")
         LOGGER.info("Step 9: Successfully downloaded object and verified checksum")
         LOGGER.info("ENDED: Verify chunk upload during server pod restart")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.tags("TEST-34261")
-    def test_server_pod_restart_kubectl_delete(self):
-        """
-        Verify IOs during pod restart (kubectl delete)
-        """
-        LOGGER.info("STARTED: Verify IOs after server pod restart (kubectl delete)")
-        del_bucket = HA_CFG["bg_bucket_ops"]["no_del_buckets"]
-        del_output = Queue()
-        event = threading.Event()  # Event to be used to send intimation of data pod deletion
-        t_t = int(perf_counter_ns())
-
-        LOGGER.info("Perform IOs with variable object sizes during server pod restart")
-        users = self.mgnt_ops.create_account_users(nusers=1)
-        self.s3_clean.update(users)
-        access_key = list(users.values())[0]["accesskey"]
-        secret_key = list(users.values())[0]["secretkey"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Step 1.1: Perform WRITEs on %s buckets for background DELETEs", del_bucket)
-        test_prefix_del = f'test-del-34261-{t_t}'
-        args = {'test_prefix': test_prefix_del, 'test_dir_path': self.test_dir_path,
-                'skipget': True, 'skipdel': True, 'bkts_to_wr': del_bucket, 'output': del_output}
-        self.ha_obj.put_get_delete(event, s3_test_obj, **args)
-        wr_resp = tuple()
-        while len(wr_resp) != 3:
-            wr_resp = del_output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-        written_bck = s3_test_obj.bucket_list()[1]
-        assert_utils.assert_equal(len(written_bck), del_bucket,
-                                  f"Failed to create {del_bucket} number of buckets."
-                                  f"Created {len(written_bck)} number of buckets")
-        LOGGER.info("Step 1.1: Performed WRITEs on %s buckets for background DELETEs", del_bucket)
-
-        LOGGER.info("Step 1.2: Perform WRITEs with variable object sizes for parallel READs")
-        test_prefix_read = f'test-read-34261-{t_t}'
-        resp = self.ha_obj.ha_s3_workload_operation(s3userinfo=list(users.values())[0],
-                                                    log_prefix=test_prefix_read, skipread=True,
-                                                    skipcleanup=True)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 1.2: Performed WRITEs with variable sizes objects for parallel READs.")
-
-        LOGGER.info("Step 2.1: Start WRITEs with variable object sizes in background")
-        test_prefix_write = f'test-write-34261-{t_t}'
-        output_wr = Queue()
-        event_set_clr = [False]
-        args = {'s3userinfo': list(users.values())[0], 'log_prefix': test_prefix_write,
-                'nclients': 2, 'nsamples': 10, 'skipread': True, 'skipcleanup': True,
-                'output': output_wr, 'setup_s3bench': False, 'event_set_clr': event_set_clr}
-        thread_wri = threading.Thread(target=self.ha_obj.event_s3_operation, args=(event,),
-                                      kwargs=args)
-        thread_wri.daemon = True  # Daemonize thread
-        LOGGER.info("Step 2.1: Successfully started WRITEs with variable sizes objects"
-                    " in background")
-
-        LOGGER.info("Step 2.2: Start READs and verify DI on the written data in background")
-        output_rd = Queue()
-        args = {'s3userinfo': list(users.values())[0], 'log_prefix': test_prefix_read,
-                'nclients': 1, 'nsamples': 10, 'skipwrite': True, 'skipcleanup': True,
-                'output': output_rd, 'setup_s3bench': False, 'event_set_clr': event_set_clr}
-        thread_rd = threading.Thread(target=self.ha_obj.event_s3_operation, args=(event,),
-                                     kwargs=args)
-        thread_rd.daemon = True  # Daemonize thread
-        LOGGER.info("Step 2.2: Successfully started READs and verify DI on the written data in "
-                    "background")
-
-        LOGGER.info("Step 2.3: Starting DELETEs of %s buckets in background", del_bucket)
-        del_buckets = written_bck.copy()
-        args = {'test_prefix': test_prefix_del, 'test_dir_path': self.test_dir_path,
-                'skipput': True, 'skipget': True, 'bkt_list': del_buckets,
-                'bkts_to_del': del_bucket, 'output': del_output}
-        thread_del = threading.Thread(target=self.ha_obj.put_get_delete,
-                                      args=(event, s3_test_obj,), kwargs=args)
-        thread_del.daemon = True  # Daemonize thread
-        thread_wri.start()
-        thread_rd.start()
-        thread_del.start()
-        LOGGER.info("Step 2.3: Started WRITEs, READs and DELETEs with variable object sizes "
-                    "during server pod restart using kubectl delete")
-        LOGGER.info("Waiting for %s seconds", HA_CFG["common_params"]["10sec_delay"])
-        time.sleep(HA_CFG["common_params"]["10sec_delay"])
-
-        LOGGER.info("Step 3: Restart the server pod by kubectl delete.")
-        LOGGER.info("Get pod name to be deleted")
-        pod_list = self.node_master_list[0].get_all_pods(pod_prefix=const.SERVER_POD_NAME_PREFIX)
-        pod_name = self.system_random.sample(pod_list, 1)[0]
-        LOGGER.info("Deleting pod %s", pod_name)
-        event.set()
-        resp = self.node_master_list[0].delete_pod(pod_name=pod_name, force=True)
-        LOGGER.debug("Response: %s", resp)
-        assert_utils.assert_true(resp[0], f"Failed to delete pod {pod_name} by kubectl delete")
-        LOGGER.info("Step 3: Successfully restarted pod %s by kubectl delete", pod_name)
-
-        LOGGER.info("Step 4: Check cluster status")
-        resp = self.ha_obj.poll_cluster_status(self.node_master_list[0])
-        assert_utils.assert_true(resp[0], resp)
-        LOGGER.info("Step 4: Cluster is in good state")
-        event.clear()
-        thread_wri.join()
-        thread_rd.join()
-        thread_del.join()
-
-        LOGGER.info("Background WRITEs, READs and DELETEs threads joined successfully.")
-        LOGGER.info("Step 5: Verify responses from background processes")
-        LOGGER.info("Step 5.1: Verify status for In-flight WRITEs")
-        responses_wr = dict()
-        while len(responses_wr) != 2:
-            responses_wr = output_wr.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-        pass_logs = list(x[1] for x in responses_wr["pass_res"])
-        fail_logs = list(x[1] for x in responses_wr["fail_res"])
-        resp = self.ha_obj.check_s3bench_log(file_paths=pass_logs)
-        assert_utils.assert_false(len(resp[1]), f"WRITEs logs which contain failures: {resp[1]}")
-        resp = self.ha_obj.check_s3bench_log(file_paths=fail_logs, pass_logs=False)
-        LOGGER.debug("WRITEs Response for fail logs: %s", resp)
-        # TODO: Uncomment following once CORTX-28541 is fixed and re-test
-        # assert_utils.assert_true(len(resp[1]) != len(fail_logs), "Some In-flight WRITEs are "
-        #                                                          "expected to fail. No failure is"
-        #                                                          f"observed in {resp[1]}")
-        LOGGER.info("Step 5.1: Verified status for In-flight WRITEs")
-
-        LOGGER.info("Step 5.2: Verifying responses from READs background process")
-        responses_rd = dict()
-        while len(responses_rd) != 2:
-            responses_rd = output_rd.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-        pass_logs = list(x[1] for x in responses_rd["pass_res"])
-        fail_logs = list(x[1] for x in responses_rd["fail_res"])
-        resp = self.ha_obj.check_s3bench_log(file_paths=pass_logs)
-        assert_utils.assert_false(len(resp[1]),
-                                  f"READs/VerifyDI logs which contain failures: {resp[1]}")
-        resp = self.ha_obj.check_s3bench_log(file_paths=fail_logs, pass_logs=False)
-        LOGGER.info("READs Response for fail logs: %s", resp[1])
-        # TODO: Uncomment following once CORTX-28541 is fixed and re-test
-        # assert_utils.assert_true(len(resp[1]) != len(fail_logs), "Some In-flight READs are "
-        #                                                          "expected to fail. No failure is"
-        #                                                          f"observed in {resp[1]}")
-
-        LOGGER.info("Step 5.2: Verified responses from READs background process")
-
-        LOGGER.info("Step 5.3: Verifying responses from DELETEs background process")
-        del_resp = tuple()
-        while len(del_resp) != 2:
-            del_resp = del_output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-        if not del_resp:
-            assert_utils.assert_true(False, "Background process failed to do DELETEs")
-        event_del_bkt = del_resp[0]  # Contains buckets when event was set
-        fail_del_bkt = del_resp[1]  # Contains buckets which failed when event was clear
-        assert_utils.assert_false(len(fail_del_bkt), "Expected pass, buckets which failed in "
-                                                     f"DELETEs: {fail_del_bkt}.")
-        # TODO: Uncomment following once CORTX-28541 is fixed and re-test
-        # assert_utils.assert_true(len(event_del_bkt), "Expected FAIL when event was set")
-        LOGGER.info("Failed buckets while in-flight DELETEs operation : %s", event_del_bkt)
-        LOGGER.info("Step 5.3: Verified responses from DELETEs background process")
-        LOGGER.info("Step 5: Verified responses from background processes")
-
-        LOGGER.info("Step 6: Start IOs (create s3 acc, buckets and upload objects) after pod "
-                    "restart by kubectl delete")
-        users = self.mgnt_ops.create_account_users(nusers=1)
-        self.s3_clean.update(users)
-        self.test_prefix = 'test-34261'
-        resp = self.ha_obj.ha_s3_workload_operation(s3userinfo=list(users.values())[0],
-                                                    log_prefix=self.test_prefix,
-                                                    setup_s3bench=False)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 6: Successfully IOs completed after pod restart by kubectl delete")
-        LOGGER.info("COMPLETED: Verify IOs after server pod restart (kubectl delete)")
